Fix country deletion error
If country with existing users and scores is about to be deleted,
deletion count will be 0 and fails the function check.
This fixes the bug.

+++ src/utils/prisma/users.ts
@@ -180,20 +180,29 @@ export async function removeUser(id: number) {
 
 export async function removeUserByCountryId(id: number) {
 	try {
-		const result = await prisma.users.deleteMany({
+		const users = await prisma.users.findMany({
 			where: {
 				countryId: id
 			}
 		});
 
-		if(result.count > 0) {
+		if(users.length > 0) {
+			const result = await prisma.users.deleteMany({
+				where: {
+					countryId: id
+				}
+			});
+
+			if(result.count <= 0) {
+				log("Invalid deleted user record.", LogLevel.ERROR);
+				return -1;
+			}
+
 			log(`users: Deleted ${ result.count } rows.`, LogLevel.LOG);
-		}
-		else {
-			log("Invalid deleted user record.", LogLevel.ERROR);
+			return result.count;
 		}
 
-		return result.count;
+		return 0;
 	}
 	catch (e) {
 		if(e instanceof Prisma.PrismaClientKnownRequestError) {
